feat: refactor access control to enable swapping implementation
