+++ libs/access-control/Cargo.toml
@@ -8,7 +8,11 @@ actix-http.workspace = true
 app-error.workspace = true
 anyhow.workspace = true
 async-trait.workspace = true
-casbin = { version = "2.2.0", features = ["cached", "runtime-tokio", "incremental"] }
+casbin = { version = "2.2.0", features = [
+    "cached",
+    "runtime-tokio",
+    "incremental",
+], optional = true }
 database.workspace = true
 database-entity.workspace = true
 futures-util.workspace = true
@@ -17,8 +21,11 @@ prometheus-client.workspace = true
 redis.workspace = true
 sqlx = { workspace = true, default-features = false, features = ["postgres"] }
 tracing.workspace = true
-tokio = { workspace = true, features = [
-    "macros",
-    "time",
-] }
+tokio = { workspace = true, features = ["macros", "time"] }
 tokio-stream.workspace = true
+uuid = { version = "1.8.0", features = ["v4"] }
+serde = { version = "1.0.200", features = ["derive"] }
+
+[features]
+default = ["casbin"]
+casbin = ["dep:casbin"]
