+++ libs/access-control/src/metrics.rs
@@ -1,10 +1,12 @@
 use prometheus_client::metrics::gauge::Gauge;
 use std::sync::atomic::{AtomicI64, Ordering};
 use std::sync::Arc;
+use std::time::Duration;
 
-use crate::enforcer::ENFORCER_METRICS_TICK_INTERVAL;
 use prometheus_client::registry::Registry;
 use tokio::time::interval;
+
+pub const ENFORCER_METRICS_TICK_INTERVAL: Duration = Duration::from_secs(120);
 
 #[derive(Clone)]
 pub struct AccessControlMetrics {
