--- libs/workspace-access/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "workspace-access"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-access-control.workspace = true
-app-error = { workspace = true, features = ["sqlx_error", "tokio_error"] }
-async-trait = "0.1.77"
-database-entity.workspace = true
-tokio = { workspace = true, features = ["sync"] }
-tracing = "0.1.40"
-uuid = { version = "1.8.0", features = ["v4"] }
-serde = { version = "1.0.200", features = ["derive"] }
