--- libs/workspace-access/src/lib.rs
@@ -1,4 +1,0 @@
-mod access_control;
-pub mod notification;
-
-pub use access_control::*;
