+++ services/appflowy-collaborate/src/state.rs
@@ -1,11 +1,11 @@
 use std::sync::Arc;
 
+use access_control::casbin::access::AccessControl;
 use dashmap::DashMap;
 use futures_util::StreamExt;
 use sqlx::PgPool;
 use uuid::Uuid;
 
-use access_control::access::AccessControl;
 use access_control::metrics::AccessControlMetrics;
 use app_error::AppError;
 use database::user::{select_all_uid_uuid, select_uid_from_uuid};
