+++ src/api/data_import.rs
@@ -87,7 +87,7 @@ async fn import_data_handler(
 
   let workspace = create_empty_workspace(
     &state.pg_pool,
-    &state.workspace_access_control,
+    state.workspace_access_control.clone(),
     &state.collab_access_control_storage,
     &user_uuid,
     uid,
