+++ src/api/workspace.rs
@@ -17,7 +17,6 @@ use tracing::{error, event, info, instrument, trace};
 use uuid::Uuid;
 use validator::Validate;
 
-use access_control::collab::CollabAccessControl;
 use app_error::AppError;
 use appflowy_collaborate::actix_ws::entities::ClientStreamMessage;
 use appflowy_collaborate::indexer::IndexerProvider;
@@ -232,7 +231,7 @@ async fn create_workspace_handler(
   let uid = state.user_cache.get_user_uid(&uuid).await?;
   let new_workspace = workspace::ops::create_workspace_for_user(
     &state.pg_pool,
-    &state.workspace_access_control,
+    state.workspace_access_control.clone(),
     &state.collab_access_control_storage,
     &uuid,
     uid,
@@ -351,7 +350,7 @@ async fn post_accept_workspace_invite_handler(
   // Currently, when the server get restarted, the policy in access control will be lost.
   workspace::ops::accept_workspace_invite(
     &state.pg_pool,
-    &state.workspace_access_control,
+    state.workspace_access_control.clone(),
     user_uid,
     &user_uuid,
     &invite_id,
@@ -369,7 +368,7 @@ async fn get_workspace_settings_handler(
   let uid = state.user_cache.get_user_uid(&user_uuid).await?;
   let settings = workspace::ops::get_workspace_settings(
     &state.pg_pool,
-    &state.workspace_access_control,
+    state.workspace_access_control.clone(),
     &workspace_id,
     &uid,
   )
@@ -389,7 +388,7 @@ async fn post_workspace_settings_handler(
   let uid = state.user_cache.get_user_uid(&user_uuid).await?;
   let settings = workspace::ops::update_workspace_settings(
     &state.pg_pool,
-    &state.workspace_access_control,
+    state.workspace_access_control.clone(),
     &workspace_id,
     &uid,
     data,
@@ -435,7 +434,7 @@ async fn remove_workspace_member_handler(
     &state.pg_pool,
     &workspace_id,
     &member_emails,
-    &state.workspace_access_control,
+    state.workspace_access_control.clone(),
   )
   .await?;
 
@@ -482,7 +481,7 @@ async fn leave_workspace_handler(
     &state.pg_pool,
     &workspace_id,
     &user_uuid,
-    &state.workspace_access_control,
+    state.workspace_access_control.clone(),
   )
   .await?;
   Ok(AppResponse::Ok().into())
@@ -508,7 +507,7 @@ async fn update_workspace_member_handler(
       &state.pg_pool,
       &workspace_id,
       &changeset,
-      &state.workspace_access_control,
+      state.workspace_access_control.clone(),
     )
     .await?;
   }
@@ -979,8 +978,12 @@ async fn add_collab_member_handler(
     );
   }
 
-  biz::collab::ops::create_collab_member(&state.pg_pool, &payload, &state.collab_access_control)
-    .await?;
+  biz::collab::ops::create_collab_member(
+    &state.pg_pool,
+    &payload,
+    state.collab_access_control.clone(),
+  )
+  .await?;
   Ok(Json(AppResponse::Ok()))
 }
 
@@ -1005,7 +1008,7 @@ async fn update_collab_member_handler(
     &state.pg_pool,
     &user_uuid,
     &payload,
-    &state.collab_access_control,
+    state.collab_access_control.clone(),
   )
   .await?;
   Ok(Json(AppResponse::Ok()))
@@ -1026,8 +1029,12 @@ async fn remove_collab_member_handler(
   state: Data<AppState>,
 ) -> Result<Json<AppResponse<()>>> {
   let payload = payload.into_inner();
-  biz::collab::ops::delete_collab_member(&state.pg_pool, &payload, &state.collab_access_control)
-    .await?;
+  biz::collab::ops::delete_collab_member(
+    &state.pg_pool,
+    &payload,
+    state.collab_access_control.clone(),
+  )
+  .await?;
 
   Ok(Json(AppResponse::Ok()))
 }
