+++ src/biz/workspace/ops.rs
@@ -64,7 +64,7 @@ pub async fn delete_workspace_for_user(
 /// object.
 pub async fn create_empty_workspace(
   pg_pool: &PgPool,
-  workspace_access_control: &impl WorkspaceAccessControl,
+  workspace_access_control: Arc<dyn WorkspaceAccessControl>,
   collab_storage: &Arc<CollabAccessControlStorage>,
   user_uuid: &Uuid,
   user_uid: i64,
@@ -117,7 +117,7 @@ pub async fn create_empty_workspace(
 
 pub async fn create_workspace_for_user(
   pg_pool: &PgPool,
-  workspace_access_control: &impl WorkspaceAccessControl,
+  workspace_access_control: Arc<dyn WorkspaceAccessControl>,
   collab_storage: &Arc<CollabAccessControlStorage>,
   user_uuid: &Uuid,
   user_uid: i64,
@@ -301,7 +301,7 @@ pub async fn open_workspace(
 
 pub async fn accept_workspace_invite(
   pg_pool: &PgPool,
-  workspace_access_control: &impl WorkspaceAccessControl,
+  workspace_access_control: Arc<dyn WorkspaceAccessControl>,
   user_uid: i64,
   user_uuid: &Uuid,
   invite_id: &Uuid,
@@ -527,7 +527,7 @@ pub async fn leave_workspace(
   pg_pool: &PgPool,
   workspace_id: &Uuid,
   user_uuid: &Uuid,
-  workspace_access_control: &impl WorkspaceAccessControl,
+  workspace_access_control: Arc<dyn WorkspaceAccessControl>,
 ) -> Result<(), AppResponseError> {
   let email = database::user::select_email_from_user_uuid(pg_pool, user_uuid).await?;
   remove_workspace_members(pg_pool, workspace_id, &[email], workspace_access_control).await
@@ -537,7 +537,7 @@ pub async fn remove_workspace_members(
   pg_pool: &PgPool,
   workspace_id: &Uuid,
   member_emails: &[String],
-  workspace_access_control: &impl WorkspaceAccessControl,
+  workspace_access_control: Arc<dyn WorkspaceAccessControl>,
 ) -> Result<(), AppResponseError> {
   let mut txn = pg_pool
     .begin()
@@ -583,7 +583,7 @@ pub async fn update_workspace_member(
   pg_pool: &PgPool,
   workspace_id: &Uuid,
   changeset: &WorkspaceMemberChangeset,
-  workspace_access_control: &impl WorkspaceAccessControl,
+  workspace_access_control: Arc<dyn WorkspaceAccessControl>,
 ) -> Result<(), AppError> {
   if let Some(role) = &changeset.role {
     upsert_workspace_member(pg_pool, workspace_id, &changeset.email, role.clone()).await?;
@@ -610,7 +610,7 @@ pub async fn get_workspace_document_total_bytes(
 
 pub async fn get_workspace_settings(
   pg_pool: &PgPool,
-  workspace_access_control: &impl WorkspaceAccessControl,
+  workspace_access_control: Arc<dyn WorkspaceAccessControl>,
   workspace_id: &Uuid,
   owner_uid: &i64,
 ) -> Result<AFWorkspaceSettings, AppResponseError> {
@@ -631,7 +631,7 @@ pub async fn get_workspace_settings(
 
 pub async fn update_workspace_settings(
   pg_pool: &PgPool,
-  workspace_access_control: &impl WorkspaceAccessControl,
+  workspace_access_control: Arc<dyn WorkspaceAccessControl>,
   workspace_id: &Uuid,
   owner_uid: &i64,
   change: AFWorkspaceSettingsChange,
