+++ src/middleware/mod.rs
@@ -1,5 +1,4 @@
 pub mod access_control_mw;
-// pub mod cors_mw;
 pub mod encrypt_mw;
 pub mod metrics_mw;
 pub mod request_id;
